+++ projects/safe/src/lib/components/widgets/editor-settings/editor-settings.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { SafeEditorSettingsComponent } from './editor-settings.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { EditorModule } from '@progress/kendo-angular-editor';
+import { EditorModule, TINYMCE_SCRIPT_SRC } from '@tinymce/tinymce-angular';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { TranslateModule } from '@ngx-translate/core';
@@ -19,5 +19,8 @@ import { TranslateModule } from '@ngx-translate/core';
     TranslateModule,
   ],
   exports: [SafeEditorSettingsComponent],
+  providers: [
+    { provide: TINYMCE_SCRIPT_SRC, useValue: 'tinymce/tinymce.min.js' },
+  ],
 })
 export class SafeEditorSettingsModule {}
