+++ projects/safe/src/lib/components/widgets/editor/editor.component.scss
@@ -2,3 +2,7 @@
   width: 100%;
   height: 100%;
 }
+
+.widget-content {
+  padding: 0.7rem 1rem;
+}
